// frontend/lib/api/generated/person/person.ts
export type postApiPersonResponse200TextPlain = {
  data: Person
  status: 200
}

export type postApiPersonResponse200ApplicationJson = {
  data: Person
  status: 200
}

export type postApiPersonResponse200TextJson = {
  data: Person
  status: 200
}

export type postApiPersonResponse201TextPlain = {
  data: Person
  status: 201
}

export type postApiPersonResponse201ApplicationJson = {
  data: Person
  status: 201
}

export type postApiPersonResponse201TextJson = {
  data: Person
  status: 201
}

export type postApiPersonResponse400TextPlain = {
  data: ProblemDetails
  status: 400
}

export type postApiPersonResponse400ApplicationJson = {
  data: ProblemDetails
  status: 400
}

export type postApiPersonResponse400TextJson = {
  data: ProblemDetails
  status: 400
}

export type postApiPersonResponse500 = {
  data: void
  status: 500
}

export type postApiPersonResponseSuccess = (postApiPersonResponse200TextPlain | postApiPersonResponse200ApplicationJson | postApiPersonResponse200TextJson | postApiPersonResponse201TextPlain | postApiPersonResponse201ApplicationJson | postApiPersonResponse201TextJson) & {
  headers: Headers;
};